--- cinesofa/cinesofa/ListaFilmes.swift
@@ -1,73 +1,0 @@
-//
-//  ListaFilmes.swift
-//  cinesofa
-//
-//  Created by Fabricio Rodrigo Baixo on 1/7/21.
-//  Copyright © 2021 Fabricio Rodrigo Baixo. All rights reserved.
-//
-
-import Foundation
-
-// MARK: - Welcome
-struct Welcome: Codable {
-    let page: Int
-    let results: [Result]
-    let totalPages, totalResults: Int
-
-    enum CodingKeys: String, CodingKey {
-        case page, results
-        case totalPages = "total_pages"
-        case totalResults = "total_results"
-    }
-}
-
-// MARK: - Result
-struct Result: Codable {
-    let voteAverage: Double
-    let title: String?
-    let overview: String
-    let releaseDate: String?
-    let adult: Bool?
-    let backdropPath: String
-    let genreIDS: [Int]
-    let voteCount: Int
-    let originalLanguage: OriginalLanguage
-    let originalTitle: String?
-    let posterPath: String
-    let id: Int
-    let video: Bool?
-    let popularity: Double
-    let mediaType: MediaType
-    let firstAirDate, name: String?
-    let originCountry: [String]?
-    let originalName: String?
-
-    enum CodingKeys: String, CodingKey {
-        case voteAverage = "vote_average"
-        case title, overview
-        case releaseDate = "release_date"
-        case adult
-        case backdropPath = "backdrop_path"
-        case genreIDS = "genre_ids"
-        case voteCount = "vote_count"
-        case originalLanguage = "original_language"
-        case originalTitle = "original_title"
-        case posterPath = "poster_path"
-        case id, video, popularity
-        case mediaType = "media_type"
-        case firstAirDate = "first_air_date"
-        case name
-        case originCountry = "origin_country"
-        case originalName = "original_name"
-    }
-}
-
-enum MediaType: String, Codable {
-    case movie = "movie"
-    case tv = "tv"
-}
-
-enum OriginalLanguage: String, Codable {
-    case en = "en"
-    case ja = "ja"
-}
